# openapi.yaml
openapi: 3.0.3
info:
  title: Payments API (Mini)
  version: 1.0.0
servers:
  - url: https://{tenant}.payments.{region}.example.com/api
    variables:
      tenant:
        default: acme
      region:
        enum:
          - us
          - eu
          - ap
        default: us
security:
  - bearerAuth: []
paths:
  /v1/payments:
    post:
      summary: Create a payment
      parameters:
        - in: header
          name: X-Idempotency-Key
          description: Client-generated key; replaying it returns the original payment.
          required: true
          schema:
            type: string
            example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /v1/payments/{paymentId}:
    get:
      summary: Get payment status
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            example: pay_1001
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Not found
  /v1/payments/{paymentId}/cancel:
    post:
      summary: Cancel payment (pre‑settlement)
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            example: pay_1001
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '409':
          description: Cannot cancel
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Payment:
      type: object
      required:
        - id
        - amount
        - currency
        - status
      properties:
        id:
          type: string
          example: pay_1001
        amount:
          type: integer
          example: 2500
        currency:
          type: string
          example: USD
        status:
          type: string
          enum:
            - PENDING
            - AUTHORIZED
            - SETTLED
            - FAILED
            - CANCELLED
        createdAt:
          type: string
          format: date-time
          example: '2026-02-23T16:00:00Z'
    PaymentCreateRequest:
      type: object
      required:
        - amount
        - currency
        - method
      properties:
        amount:
          type: integer
          example: 2500
        currency:
          type: string
          example: USD
        method:
          $ref: '#/components/schemas/PaymentMethod'
        description:
          type: string
          example: Order 4821
    PaymentMethod:
      type: object
      properties:
        type:
          type: string
          enum:
            - card
        cardToken:
          type: string
          example: tok_visa_4242